Nova to honor "cross_az_attach" during server(VM) migrations
While querying placement to discover hosts to execute virtual
machines (VMs) migration (or resize, which is a cold VM migration),
Nova is not sending the current availability zone (AZ) of the VM.
Therefore, there are situations where nova decides to migrate the VM
to a host in a different AZ, which causes problems if the setup is
using `cross_az_attach=False`.

Other things such as server groups, affinity and antiaffinity rules
are honored in migrate and resize. Therefore, it makes sense to take
the situation with the "cross_az_attach" option as well.

By querying Placement with the current AZ of the VM, we (Nova) will
receive only hosts in the given AZ, which in turn resolves the
mentioned issue of cross zone migration during VMs resize or host
migration.

Change-Id: Id6c5d9bed4ea34c1f1b083264d7cc47ad941fbee